requirements.txt
asgiref==3.5.2
attrs==22.1.0
Babel==2.11.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
Django==3.2.16
django-extra-views==0.13.0
django-haystack==3.2.1
django-oscar==3.1
django-phonenumber-field==5.2.0
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.12
environ==1.0
exceptiongroup==1.0.1
factory-boy==3.2.1
Faker==15.2.0
idna==3.4
iniconfig==1.1.1
npm==0.1.1
optional-django==0.1.0
packaging==21.3
phonenumbers==8.12.57
Pillow==9.3.0
pluggy==1.0.0
purl==1.6
pycountry==22.3.5
pyparsing==3.0.9
pysolr==3.9.0
pytest==7.2.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
sorl-thumbnail==12.6.3
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.12

cffi~=1.15.0
future~=0.18.2
pip~=22.0.4
cryptography~=36.0.1
tornado~=6.1
Jinja2~=3.1.1
setuptools~=57.4.0
psycopg2~=2.9.3
PyYAML~=6.0
py~=1.11.0
chardet~=4.0.0
tzdata~=2021.5
djangorestframework~=3.13.1
django-filter~=21.1